+++ src/main/java/insane96mcp/enhancedai/setup/Strings.java
@@ -5,6 +5,7 @@ import insane96mcp.enhancedai.EnhancedAI;
 public class Strings {
 	public static class Tags {
 		public static final String LAUNCH = EnhancedAI.RESOURCE_PREFIX + "launch";
+		public static final String BREACH = EnhancedAI.RESOURCE_PREFIX + "breach";
 		public static final String JOHN_CENA = EnhancedAI.RESOURCE_PREFIX + "john_cena";
 	}
 }