+++ changelog.md
@@ -5,6 +5,7 @@
 * Slowness is no longer applied from thrown webs if the player parrys the projectile
 * Fixed spiders throwing webs even they could not see the player
 * Fixed skeletons running from target having faster speed when far away instead of when near. Also made those speeds configurable.
+* Fixed skeletons not strafing at all
 
 ## Alpha 1.2.1
 * Fixed server startup crash