+++ changelog.md
@@ -5,6 +5,7 @@
 * Spider Cobweb Throwing
   * Cobweb damage is now configurable and scales with difficulty
   * Cobweb cooldown is now configurable
+  * Min and max attack distance is now configurable
   * Cobwebs placed will now disappear after 5 seconds (configurable)
 * Slowness is no longer applied from thrown webs if the player parrys the projectile
 * Fixed spiders throwing webs even they could not see the player